// src/lib/interfaces/WeatherAPI.ts
export interface Location {
    lat: number;
    lon: number;
    message?: string;
}

export interface WeatherData {
    coord: {
        lon: number;
        lat: number;
    };
    weather: Array<
        {
            id: number;
            main: string;
            description: string;
            icon: string;
        }
    >;
    base: string;
    main: {
        temp: number;
        feels_like: number;
        temp_min: number;
        temp_max: number;
        pressure: number;
        humidity: number;
        sea_level: number;
        grnd_level: number;
    };
    visibility: number;
    wind: {
        speed: number;
        deg: number;
        gust?: number;
    };
    rain?: {
        "1h"?: number;
        "3h"?: number;
    };
    snow?: {
        "1h"?: number;
        "3h"?: number;
    };
    clouds: {
        all: number;
    };
    dt: number;
    sys: {
        type: number;
        id: number;
        country: string;
        sunrise: number;
        sunset: number;
    };
    timezone: number;
    id: number;
    name: string;
    cod: number;
}

export interface HomePageTodayData {
    weatherData: {
        city: string;
        countryCode: string;
        temp: number;
        temp_min: number;
        temp_max: number;
        humidity: number;
        weather: string;
        description: string;
        icon: string;
    };
    countryName: string;
    icon: string;
    hour: number;
    minute: number;
}